Change animation from fade out effect to slide up + fade out
I had a hard time getting this to work. At first, I initialised the style as
```
Animation.style
    [ Animation.translate3d (Animation.percent 0) (Animation.percent 0) (Animation.percent 0)
    , Animation.opacity 1.0
    ]
```
And updated it with
```
Animation.to
    [ Animation.translate3d (Animation.percent 0) (Animation.percent -100) (Animation.percent 0)
    , Animation.opacity 0
    ]
```
but while the opacity change was working (and visible in the style tag), nothing was happening with `translate3d`.

It turns out that I needed to use pixel values, except for the value I wanted to change. e.g. initialise as
```
Animation.style
    [ Animation.translate3d (Animation.px 0) (Animation.px 0) (Animation.px 0)
    , Animation.opacity 1.0
    ]
```
And update as
```
Animation.to
    [ Animation.translate3d (Animation.px 0) (Animation.percent -100) (Animation.px 0)
    , Animation.opacity 0
    ]
```

Thanks to akoppela (https://github.com/akoppela) on the Elm Slack for helping me figure out the problem.