Change animation: Slide up effect
Or rather, the item appears to shrink away up to its top. This is achieved by using `scale3d` and `transform-origin`.

Setting Animation.scale3d to 0 0 0 means that the item shrinks inward to its centre. As a corollary, setting 1 0 0 means that the x-axis value is maintained, so the item shrinks to its vertical middle.

Using Animation.transformOrigin and setting its first value (the y-offset) to -50% is what makes item shrink up towards its top.